src/main/java/uk/gov/hmcts/divorce/systemupdate/event/AdminUnlinkApplicant1FromCase.java
package uk.gov.hmcts.divorce.systemupdate.event;

import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;
import uk.gov.hmcts.ccd.sdk.api.CCDConfig;
import uk.gov.hmcts.ccd.sdk.api.CaseDetails;
import uk.gov.hmcts.ccd.sdk.api.ConfigBuilder;
import uk.gov.hmcts.ccd.sdk.api.callback.AboutToStartOrSubmitResponse;
import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
import uk.gov.hmcts.divorce.divorcecase.model.State;
import uk.gov.hmcts.divorce.divorcecase.model.UserRole;
import uk.gov.hmcts.divorce.solicitor.service.CcdAccessService;

import java.util.List;

import static uk.gov.hmcts.divorce.divorcecase.model.State.STATES_NOT_WITHDRAWN_OR_REJECTED;
import static uk.gov.hmcts.divorce.divorcecase.model.UserRole.CREATOR;
import static uk.gov.hmcts.divorce.divorcecase.model.UserRole.SUPER_USER;
import static uk.gov.hmcts.divorce.divorcecase.model.access.Permissions.CREATE_READ_UPDATE;

@Slf4j
@Component
public class AdminUnlinkApplicant1FromCase implements CCDConfig<CaseData, State, UserRole> {

    @Autowired
    private CcdAccessService ccdAccessService;

    public static final String ADMIN_UNLINK_APPLICANT_1 = "admin-unlink-applicant1";

    @Override
    public void configure(final ConfigBuilder<CaseData, State, UserRole> configBuilder) {
        if (Boolean.parseBoolean(System.getenv().get("ADMIN_UNLINK_APPLICANT_1_ENABLED"))) {
            configBuilder
                .event(ADMIN_UNLINK_APPLICANT_1)
                .forStates(STATES_NOT_WITHDRAWN_OR_REJECTED)
                .name("Unlink Applicant1 from case")
                .grant(CREATE_READ_UPDATE, SUPER_USER)
                .retries(120, 120)
                .aboutToSubmitCallback(this::aboutToSubmit);
        }
    }

    public AboutToStartOrSubmitResponse<CaseData, State> aboutToSubmit(CaseDetails<CaseData, State> details,
                                                                       CaseDetails<CaseData, State> beforeDetails) {

        CaseData caseData = details.getData();


        log.info("Admin unlinking Creator from case (id: {})",  details.getId());
        ccdAccessService.removeUsersWithRole(details.getId(), List.of(CREATOR.getRole()));

        return AboutToStartOrSubmitResponse.<CaseData, State>builder()
            .data(caseData)
            .build();
    }
}
